fix(tests): refactor unit tests
- refactor test constants to new file
- remove cast to 'any'
