+++ src/libs/tmdk/package.json
@@ -27,6 +27,7 @@
     "@aws-sdk/client-iottwinmaker": "^3.142.0",
     "@aws-sdk/client-kinesis-video": "^3.142.0",
     "@aws-sdk/client-s3": "^3.142.0",
+    "aws-sdk": "^2.1332.0",
     "aws-cdk": "^2.50.0",
     "aws-cdk-lib": "^2.50.0",
     "aws-sdk-mock": "^5.8.0",
